+++ src/network/slirp/misc.c
@@ -129,8 +129,8 @@ getouraddr()
 //what?!
 
 struct quehead_32 {
-	u_int32_t qh_link;
-	u_int32_t qh_rlink;
+	uintptr_t qh_link;
+	uintptr_t qh_rlink;
 };
 
 inline void
@@ -141,10 +141,10 @@ insque_32(a, b)
 	register struct quehead_32 *element = (struct quehead_32 *) a;
 	register struct quehead_32 *head = (struct quehead_32 *) b;
 	element->qh_link = head->qh_link;
-	head->qh_link = (u_int32_t)element;
-	element->qh_rlink = (u_int32_t)head;
+	head->qh_link = (uintptr_t)element;
+	element->qh_rlink = (uintptr_t)head;
 	((struct quehead_32 *)(element->qh_link))->qh_rlink
-	= (u_int32_t)element;
+	= (uintptr_t)element;
 }
 
 inline void
